refactor(server): external_role_conditions を oauth/internal.ts に移動
provider_id (GitHub/Discord) を軸にした設定であり、
oauthConnections と同じ文脈に属するため app.ts から移動。
